Conflicts:
	bindings/py/setup.py
	src/test/unit/algorithms/SpatialPoolerTest.cpp

+++ src/nupic/proto/ClaClassifier.capnp
@@ -5,8 +5,8 @@ using import "/nupic/proto/BitHistory.capnp".BitHistoryProto;
 # Next ID: 15
 struct ClaClassifierProto {
   steps @0 :List(UInt16);
-  alpha @1 :Float32;
-  actValueAlpha @2 :Float32;
+  alpha @1 :Float64;
+  actValueAlpha @2 :Float64;
   learnIteration @3 :UInt32;
   recordNumMinusLearnIteration @4 :UInt32;
   recordNumMinusLearnIterationSet @5 :Bool;
@@ -15,7 +15,7 @@ struct ClaClassifierProto {
   iterationNumHistory @8 :List(UInt32);
   activeBitHistory @9 :List(StepBitHistories);
   maxBucketIdx @10 :UInt32;
-  actualValues @11 :List(Float32);
+  actualValues @11 :List(Float64);
   # Each index is true if the corresponding index in actualValues has been set.
   actualValuesSet @12 :List(Bool);
   version @13 :UInt16;
